ST-BHJ release linux

+++ linux/src/BoxCollider2d.hpp
@@ -1,4 +1,5 @@
 // 
+// Bullet Hell Jam 2022
 // SchumakerTeam
 // Hudson Schumaker
 //
@@ -15,7 +16,7 @@ class BoxCollider2d final {
     public:
         BoxCollider2d();
         BoxCollider2d(short w, short h);
-        BoxCollider2d(short x, short y, short w, short h);
+        BoxCollider2d(float x, float y, short w, short h);
         BoxCollider2d(Vector2d o, Dimension s);
 
         static bool collide(const BoxCollider2d& a, const BoxCollider2d& b);
@@ -23,8 +24,8 @@ class BoxCollider2d final {
         BoxCollider2d getBounds();
         Vector2d getLocation();
         Dimension getSize();
-        short getX() const;
-        short getY() const;
+        float getX() const;
+        float getY() const;
         short getWidth() const;
         short getHeight() const;
 };
